fix: Opcode snippet uses 32bytes for addr instead of 20 (#65)
The Mnemonic Example in lines 76-94 discusses pushing [dead_addr] and [coffee_addr] to the stack followed by the SSTORE opcode, but the values pushed are 32 bytes long (using PUSH32 opcode). This conflicts with the names given to them (addr implies a label of address type).

These values should instead be 20 bytes long and make use of the PUSH20 opcode, since 20 bytes is the correct byte length for EVM address types. I have updated the opcodes used and the values of the stack as well as the resulting value in storage shown by comments in lines 86-88.