Huff calldataload should account for func selector
The Huff code snippets detailing the addTwo() function logic currently invoke the calldataload opcode on calldata offsets 0x00 and 0x20, but that cau6ses the contract function to use the function selector when adding two uints.

Instead, the Huff code snippets should invoke the calldataload opcode on calldata offsets 0x04 and 0x24 so that the function selector is skipped over and the uints provided in the calldata of the function call are added without pollution from the function selector. I have updated all references to this bug in the provided Huff code, EVM bytecode, and line-by-line execution explanation from the values '0x00' and '0x20' to '0x04' and '0x24'

This way, the huff code compiles and the contract executes its logic as intended.